virtio_sys: move vhost ioctls to a Linux-only file
The vhost kernel interface is specific to Linux; factor it out to a
separate file so it can be more easily controlled with a cfg check.

The rest of the virtio_sys bindings are generated from Linux uapi
headers, but they are valid for all platforms, so update the file-level
doc comment to match.

BUG=None
TEST=tools/dev_container tools/presubmit
TEST=cargo build --target=aarch64-apple-darwin -p virtio_sys

Change-Id: I1c7e1b6816993499ebf4431095e9308f7611a4d4
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5005276
Reviewed-by: Cody Schuffelen <schuffelen@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>