crosvm: Add vCPU thread nice adjustment support
Adds support for specifying a nice value to be used for vCPU threads.
This can be used to boost (or deboost) the priority of the vCPU threads
relative to everything else. To boost the threads priority, the
appropriate limits will need to be set.

BUG=b:310707244
TEST=Verified nice values on "crosvm_vcpuN" using top on host.

Change-Id: Ia1ad8d2c8f88bdd411551f3be72c7eb3553417c8
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5027655
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Isaac Bosompem <mrisaacb@google.com>