devices: virtio: scsi: Remove the limit on the transfer size.
Currently crosvm imposes a limit on the transfer size. This has led to a
limited throughput on heavy workloads such as vm.Fio.scsi_stress_rw.
This commit removes the limit so that the guest driver can queue as much
requests as it wants to the virtqueue, leading to a slightly moderated
performance.

BUG=b:300042376
TEST=./tools/presubmit

Change-Id: I8c1a465e2520441b7467053e81e0796a6535fa7f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5036061
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Takaya Saeki <takayas@chromium.org>
Commit-Queue: Joe Hattori <hattorij@google.com>