vmm_vhost: eliminate locking in MasterReqHandler and merge traits
Only `Slave` implemented `VhostUserMasterReqHandler` direcly and it was
internally using a mutex to guard everything just as done by
`VhostUserMasterReqHandlerMut` implicitly. To simplify things, the trait
now only supports `&mut self` methods and users of the trait need to add
a mutex themselves as needed.

There are two users of `VhostUserMasterReqHandler`:

1. `VhostBackendReqConnection`: Uses an `Arc` for memory management and
   so does need the lock to get a `&mut`. Added a `Mutex` directly to
   it (effectively at the same level of granularity as before).
2. `MasterReqHandler`: Already has a `&mut` (it can only process one
   request at a time) and so doesn't need a lock.

`Slave` no longer needs a lock internally. That will be cleaned up in a
separate CL.

Change-Id: Icf8edd40087b68e7d2b55a0569ffde1afde6bfcf
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5076926
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>