devices: vsock: add connection logging for Windows.
We're trying to identify a stall in the vsock device that appears to
occur when handling a connection request. It's not entirely clear where
or why we're stalling in the codepath, so this CL adds a bunch of
logging to help us narrow things down. It also normalizes some of the
formatting for vsock info & error messages to increase log readability.

BUG=b:313558198
TEST=ran emulator w/ guest that uses vsock.

Change-Id: I89e50a1949f160ec152344e870e1e43dcf8f2e79
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5079245
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Noah Gold <nkgold@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Richard Zhang <rizhang@google.com>