rutabaga_gfx: decouple minigbm build from virglrenderer
The current build.rs assumes that minigbm is only a dependency of
virglrenderer, such that it will not be internally cloned, built, and
statically linked if `--features=virgl_renderer` is not enabled.

This is a problem because virtio-wl and virtio-gpu cross-domain may
both make use of minigbm without also requiring virglrenderer.

Now minigbm is an orthogonal build step and occurs if at least one of
those use-cases are enabled.

BUG=b:315017032,b:265199561
TEST=cargo build crosvm
TEST=CROSVM_USE_SYSTEM_MINIGBM=1 cargo build crosvm
TEST=CROSVM_USE_SYSTEM_MINIGBM=1 CROSVM_USE_SYSTEM_VIRGLRENDERER=1 cargo build crosvm

Change-Id: I4550c3497d33bd16a0c34d354b3a973756f160f4
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5085560
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Ryan Neph <ryanneph@google.com>
Reviewed-by: Yiwei Zhang <zzyiwei@chromium.org>