crosvm: add "writable" option for create_composite
Currently, create_composite generates a read-only composite disk by default.

But in embedded system, sometime we need to use a real disk partition as the
backing file. In this case, it is required to create a writable composite disk.
So an optional "writable" argument is added in the command line.

BUG=b:318862474
TEST=cargo test --features composite-disk
TEST=./tools/presubmit
TEST=create a writable composite disk and use it as the backing of virtio-blk,
     the writing works well.

Signed-off-by: Junnan Wu <junnan01.wu@samsung.com>
Change-Id: I05033473cf28628887926b18fe17429e267f26f3
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5179893
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>