cros_async: Add AsyncFd to Tokio linux implementation
This uses tokio's utilities to re-schedule IO that would fail with a
"would block" error. The `IoSource` is chosen by the Executor based
on whether the file descriptor can be successfully added to the internal
epoll scheduler.

Specifically, AsyncFd::async_io is used for system calls that can fail
with EWOULDBLOCK, and spawn_blocking is used for system calls that will
always block.

Bug: b/320603688
Test: tools/presubmit
Test: cargo build --features=cros_async/tokio --target=x86_64-unknown-linux-gnu
Change-Id: Ibafbbe65d67cb88b5fcd0f18b00f504430c785c6
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5174967
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Cody Schuffelen <schuffelen@google.com>
Reviewed-by: Noah Gold <nkgold@google.com>