config: enable key=value options for touch input devices
This converts the parser for --multi-touch, --single-touch, and
--trackpad to allow the use of the same key=value parser that the
majority of the other crosvm command-line options use. The original
colon-separated format is also still supported as a fallback until
downstream users can migrate to the new form.

For example, an option like:

  crosvm run --multi-touch /some/socket:1024:768 ...

... would now be written as:

  crosvm run --multi-touch /some/socket,width=1024,height=768

This makes these options consistent with the rest of the crosvm CLI and
also allows the fields to contain colons, which is especially useful on
Windows, where paths often contain a colon as part of a drive specifier.

BUG=b:320385081

Change-Id: I0d82db5f17fe8664a84a86ad88d89733e934b7a1
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5207533
Reviewed-by: Jorge Moreira <jemoreira@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Noah Gold <nkgold@google.com>