devices: balloon: remove sys specific code
Always send VmMemoryClient messages and leave it up the hypervisor
implementation to ignore them if they aren't relevant.

This will result in more IPC on some hypervisors, like KVM. If it turns
out to be a problem, we can probably add a special result message that
instructs VmMemoryClient to not send futher requests.

Change-Id: I54064bf89431585f66391126aab68592fbbd2c05
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5251787
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Frederick Mayle <fmayle@google.com>