base: macos: add pipe() implementation
We can't use pipe2() since that is a Linux-only API, so O_CLOEXEC must
be applied as a separate step. This should not be a problem in practice
for crosvm, since we should never be calling pipe() simultaneously with
fork()/exec() in another thread, since fork() in a multithreaded process
is generally unsafe and our fork wrapper disallows it.

Change-Id: I27886b6e50f8b0eb018cf38161530fa7f92f4761
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5251101
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>