devices: xhci: reintroduce busy handler flag check
This CL reverts a change from https://crrev.com/c/2202745 where we had
to remove a busy check due to a bug in our implementation. The fix to
the bug is to make sure we set both EHB and Event Ring Dequeue pointer in
the same function so we don't call interrupts multiple times on the same
event.

BUG=b:309889022
TEST=`while adb shell echo hi; do : ; done` and old crash doesn't happen

Change-Id: I26f06774238fc0f40f1363012e8c41bab0d50e56
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5294354
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Auto-Submit: Morg <morg@google.com>