linux: remove memfd panic handler
This is only intended to be used on the downstream ChromeOS version of
crosvm, so we should really be patching it there rather than carrying
this code in upstream crosvm.

The crosvm version of the memfd panic handler will be replaced with the
version from libchromeos-rs.

BUG=b:325081271
BUG=b:309651697

Change-Id: I0587bacd103a3f2471d2ac7227903df6cd7e0028
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5293235
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>