cros_async: Move `concurrency` to Overlapped enum
`Executor::with_kind_and_concurrency` exists only for the benefit of the
`ExecutorKindSys::Overlapped` enum variant, it's not used outside of
windows code and not used for `ExecutorKindSys::Handle`.

`DiskOption`'s `io_concurrency` value is used to override the
concurrency setting in the `ExecutorKind`.

Test: tools/presubmit
Change-Id: Idd59d78283038ab6f986953f846cd70ef25d4324
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5315960
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Cody Schuffelen <schuffelen@google.com>
Reviewed-by: Noah Gold <nkgold@google.com>