crosvm: Delay hotplug until PCIe slot is not busy
If a PCI device is hot plugged into a PCIe slot with pending hot unplug
events unhandled by the guest OS, the new device may not be discovered
by the guest. Therefore, crosvm shall not reuse the PCIe slots until
guest indicates the previous events are handled. This is achieved by
  Prioritize vacant ports if available.
  If all ports are busy, wait for command completion asynchornously
  before sending hotplug signal.
An e2etest for the edge case is added.

BUG=b:325997618
TEST=./tools/dev_container ./tools/presubmit
TEST=tast run ${DUT} network.ARCIPv6Connectivity (downstream chromeOS on
a DUT with 2 ethernet adapters, ref: b:324979531)

Change-Id: I66d4ee9c88dd6c703491c1af440bc7160e8ff91b
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5316598
Commit-Queue: Ningyuan Wang <ningyuan@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>