usb: Implement Fido Device and OUT Endpoint
This CL introduces the implementation for a Fido Device. It is a virtual
implementation of a fido security key, taking care of operations like
transaction management (handled by a TransactionManager struct), and
read/write polling on a host hidraw fd.

It also introduces the logic for OUT endpoint transfers, where the guest
can send write requests to the host security key. IN endpoint transfers
are more complex and will be implemented in a follow-up CL.

BUG=b:297957754
TEST=presubmit and unit tests

Change-Id: Ifd19d07602c6a28c135daea063bff3107b7b9592
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5316379
Reviewed-by: Ryuichiro Chiba <chibar@chromium.org>